# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tts/speech:
    post:
      operationId: ttsSpeech
      tags: [tts]
      summary: OpenAI text-to-speech proxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TtsSpeechBody"
      responses:
        "200":
          description: Audio
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    TtsSpeechBody:
      type: object
      properties:
        text: { type: string, minLength: 1 }
        voice: { type: string }
        speed:
          type: number
          minimum: 0.25
          maximum: 4
        model:
          type: string
          enum: [tts-1, tts-1-hd, gpt-4o-mini-tts]
      required: [text]
...